DSEGOG-104 extract parser to its own file & add comments
